+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3 (pending)
+
+* Use lower case 'class' only in the exception's dimensions.
+* Deprecate Action and Controller.
+
 # 0.2
 
 * Switch to v3 api